Move pieces of Collection data into a "metadata" element:name, license, handle
