+++ dspace-rest/src/main/java/org/dspace/rest/common/CollectionMetadata.java
@@ -1,0 +1,54 @@
+package org.dspace.rest.common;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlType;
+
+/**
+ * Created with IntelliJ IDEA.
+ * User: peterdietz
+ * Date: 5/22/13
+ * Time: 9:41 AM
+ * To change this template use File | Settings | File Templates.
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "collection-metadata")
+public class CollectionMetadata {
+    //Internal metadata
+    @XmlElement(required = true)
+    private String name;
+    private String handle;
+    private String license;
+
+    //unused-metadata
+    //String provenance_description;
+    //String short_description;
+    //String introductory_text;
+    //String copyright_text;
+    //String side_bar_text;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getHandle() {
+        return handle;
+    }
+
+    public void setHandle(String handle) {
+        this.handle = handle;
+    }
+
+    public String getLicense() {
+        return license;
+    }
+
+    public void setLicense(String license) {
+        this.license = license;
+    }
+}
